% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_diag.R
\name{lm_diag}
\alias{lm_diag}
\title{Diagnose Linear Model Assumptions}
\usage{
lm_diag(model, output.p = FALSE, alpha = 0.05, ...)
}
\arguments{
\item{model}{an object of class 'lm'}

\item{output.p}{logical that specify output information type. If FALSE outputs 'pass' when failed to reject both normality and equal variance null hypotheses, or outputs 'fail', otherwise. If TRUE outputs a list containing two p-values.}

\item{alpha}{numeric alpha level for determining whether null hypotheses are rejected. Applicable only if output.p is FALSE.}

\item{...}{additional arguments to be passed to the residuals function}
}
\value{
Character string of test result, 'pass' or 'fail', or list of p-values depending on the \code{output.p} argument.
}
\description{
Diagnose linear model assumptions by testing model residuals for normality using Shapiro-Wilk test, and for homogeneity using Breusch-Pagan test.
}
\examples{
lm_model <- lm(lifeExp ~ gdpPercap, data = gapminder::gapminder)
lm_diag(lm_model)


}
